# openapi.yaml
openapi: 3.1.0
info:
  title: Kitsu DeFi API
  description: |
    Gamified DeFi savings app on TON blockchain.
    
    ## Authentication
    Most endpoints require JWT authentication. Get a token by connecting your wallet via `/auth/connect`.
    
    Include the token in the `Authorization` header as `Bearer <token>`.
  version: 1.0.0
  contact:
    name: Kitsu DeFi
    url: https://kitsu.fi
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.kitsu.fi
    description: Production
...
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me:
    patch:
      tags: [Users]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
  /transactions/deposit:
    post:
      tags: [Transactions]
      summary: Initiate deposit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Transaction initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
  /transactions/withdraw:
    post:
      tags: [Transactions]
      summary: Initiate withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: Transaction initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
  /transactions/stake:
    post:
      tags: [Transactions]
      summary: Stake via TONStakers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stake transaction initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
  /transactions/unstake:
    post:
      tags: [Transactions]
      summary: Unstake via TONStakers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unstake transaction initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
  /transactions/history:
    get:
      tags: [Transactions]
      summary: Transaction history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/connect
  schemas:
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request body

    # Health
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        walletAddr:
          type: string
          description: TON wallet address
        username:
          type: string
          nullable: true
        tonHandle:
          type: string
          nullable: true
        xp:
          type: integer
          description: Total XP points
        level:
          type: integer
          description: Current level
        streakDays:
          type: integer
          description: Consecutive days saved
        createdAt:
          type: string
          format: date-time
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
...
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        tonHandle:
          type: string
          maxLength: 128
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [deposit, withdraw, stake, unstake]
        amount:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        txHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    TransactionResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                transaction:
                  $ref: '#/components/schemas/Transaction'
    TransactionsResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                transactions:
                  type: array
                  items:
                    $ref: '#/components/schemas/Transaction'
    DepositRequest:
      type: object
      required: [type, targetId, amountTon]
      properties:
        type:
          type: string
          enum: [vault, goal, den]
        targetId:
          type: string
          format: uuid
        amountTon:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'
    WithdrawRequest:
      type: object
      required: [type, sourceId, amountTon]
      properties:
        type:
          type: string
          enum: [vault, goal, den]
        sourceId:
          type: string
          format: uuid
        amountTon:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'

    # Quest
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...